lab/embedding/poisoned_data_demo/pyproject.toml
[tool.poetry]
name = "poisoned-data-demo"
version = "0.1.0"
description = ""
authors = ["jchauhan <jitendra.chauhan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.2.4"
pandas = "^2.2.3"
umap-learn = "^0.5.7"
matplotlib = "^3.10.1"
seaborn = "^0.13.2"
scikit-learn = "^1.6.1"
sentence-transformers = "^4.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
